fixed the page routing for findings
