+++ frontend/src/components/FindingOverviewTable.vue
@@ -45,10 +45,10 @@
       </b-table-column>
       <b-table-column custom-key="actions" cell-class="is-actions-cell" v-slot="props">
         <div class="buttons is-right">
-          <router-link :to="{name:'finding.edit', params: {id: props.row.id}}" class="button is-small is-primary" v-b-tooltip.hover title="View Event">
+          <router-link :to="{name:'findings.view', params: {id: props.row.id}}" class="button is-small is-primary" v-b-tooltip.hover title="View Finding">
             <b-icon icon="information" size="is-small"/>
           </router-link>
-          <button class="button is-small is-danger" type="button" @click.prevent="trashModal(props.row)" v-b-tooltip.hover title="Delete Event">
+          <button class="button is-small is-danger" type="button" @click.prevent="trashModal(props.row)" v-b-tooltip.hover title="Archive Finding">
             <b-icon icon="trash-can" size="is-small"/>
           </button>
         </div>
